fix bug with migrations
