+++ backend/alembic/versions/45904853ac45_.py
@@ -1,0 +1,32 @@
+"""empty message
+
+Revision ID: 45904853ac45
+Revises: d8e8bcefa876
+Create Date: 2026-05-23 17:33:35.254337
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '45904853ac45'
+down_revision: Union[str, Sequence[str], None] = 'd8e8bcefa876'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    pass
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    pass
+    # ### end Alembic commands ###
